supporting docs from modifiedlardocs
